requirements.txt
streamlit
mistralai
supabase
python-dotenv
openpyxl
pymupdf
